Run and save benchmark
