+++ .github/workflows/benchmark.yml
@@ -1,0 +1,36 @@
+name: benchmark
+
+on:
+  pull_request:
+    paths-ignore:
+      - '**.md'
+
+env:
+  DEBUG: "1"
+  SHOPIFY_CLI_ENV: development
+  SHOPIFY_CONFIG: debug
+  YARN_VERSION: "1.22.19"
+  PNPM_VERSION: "7.18.1"
+
+jobs:
+  benchmark:
+    name: Node ${{ matrix.node }} in ${{ matrix.os }}
+    runs-on: ${{ matrix.os }}
+    if: ${{ github.event_name == 'push' }}
+    timeout-minutes: 30
+    strategy:
+      matrix:
+        os: ['ubuntu-latest', 'windows-latest', 'macos-latest']
+        node: ['14.20.0', '16.17.0', '18.7.0']
+    steps:
+      - uses: actions/checkout@v2
+        name: Checkout [main]
+        with:
+          fetch-depth: 0
+      - name: Setup deps
+        uses: ./.github/actions/setup-cli-deps
+        with:
+          node-version: ${{ matrix.node }}
+      - name: Build
+        working-directory: workspace
+        run: pnpm run benchmark
