+++ .github/workflows/shopify-cli.yml
@@ -106,6 +106,11 @@ jobs:
         if: ${{ matrix.os == 'ubuntu-latest' && matrix.node == '18.7.0' }}
         with:
           branch-name: main
+      - name: Run and save benchmark
+        uses: ./.github/actions/run-and-save-benchmark
+        if: ${{ matrix.os == 'ubuntu-latest' && matrix.node == '18.7.0' }}
+        with:
+          branch-name: main
 
   pr-platform-agnostic:
     name: ${{ matrix.target }} with Node ${{ matrix.node }} in ${{ matrix.os }}
@@ -187,6 +192,11 @@ jobs:
         if: ${{ matrix.os == 'ubuntu-latest' && matrix.node == '18.7.0' }}
         with:
           base-branch-name: "${{ github.base_ref }}"
+      - name: Run and save benchmark
+        uses: ./.github/actions/run-and-save-benchmark
+        if: ${{ matrix.os == 'ubuntu-latest' && matrix.node == '18.7.0' }}
+        with:
+          branch-name: "${{ github.head_ref }}"
 
   manually-triggered:
     name: Testing with Node ${{ inputs.node-version }} in ${{ inputs.os }}
